+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.3] - 2025-06-09
+
+### Fixed
+- Fixed issue where a final update was not happening for parent uuids. (DR-3678)
+
 ## [3.2.2] - 2025-04-30
 
 ### Added